♻️ refactor: organize source directories
