+++ src/index.ts
@@ -1,4 +1,4 @@
-export { Button } from './Button';
-export { Input } from './Input';
-export { Theme } from './Theme';
-export { Typography } from './Typography';
+export { Button } from './components/Button';
+export { Input } from './components/Input';
+export { Theme } from './components/Theme';
+export { Typography } from './components/Typography';
